Dedup of some common functionality, minor PR fixes
Ticket: BGX-44
