+++ src/coin/baseCoin/iface.ts
@@ -3,6 +3,50 @@ export interface BaseKey {
 }
 
 /**
+ * A private key in extended or raw format
+ */
+export type PrivateKey = {
+  prv: string;
+};
+
+/**
+ * A public key in extended, compressed, or uncompressed format
+ */
+export type PublicKey = {
+  pub: string;
+};
+
+/**
+ * A seed to create key pairs. Must be between 16 and 64 Bytes long
+ */
+export type Seed = {
+  seed: Buffer;
+};
+
+export type KeyPairOptions = Seed | PrivateKey | PublicKey;
+
+/**
+ * @param source
+ */
+export function isSeed(source: KeyPairOptions): source is Seed {
+  return (source as Seed).seed !== undefined;
+}
+
+/**
+ * @param source
+ */
+export function isPrivateKey(source: KeyPairOptions): source is PrivateKey {
+  return (source as PrivateKey).prv !== undefined;
+}
+
+/**
+ * @param source
+ */
+export function isPublicKey(source: KeyPairOptions): source is PublicKey {
+  return (source as PublicKey).pub !== undefined;
+}
+
+/**
  * Key pair in the protocol default format.
  */
 export type DefaultKeys = {
