--- src/coin/eth/enum.ts
@@ -1,13 +1,0 @@
-/**
- * Contract type describes a class of contract in a Transaction. This structure is used to identify what the Contract is.
- */
-export enum ContractType {
-  /**
-   * This is a transfer of TRX contract.
-   */
-  Send,
-  /**
-   * This is the multi-sig initialization contract.
-   */
-  WalletInitialization,
-}
